refactor: update import paths for MockFactory and remove unused registration test file
